fix(react): restore scroll area runtime test coverage
